Correção de erros da implementação em 74b690.
Deixei comentários para o que falta fazer.

+++ include/types.h
@@ -56,7 +56,7 @@
 
     struct BootBlock bootBlock;
 
-    int* indexMFT;
+    BYTE* indexMFT;
 
     struct descritor LDAA[MAX_FILES_OPEN];
   };
